+++ uoshardware/devices.py
@@ -1,41 +1,11 @@
 """Module defining supported hardware definitions."""
+from dataclasses import dataclass
+
 from uoshardware import Persistence
 from uoshardware.abstractions import Device, Pin, UOSFunctions
 from uoshardware.interface import Interface
 
-
-# This is an interface for client implementations dead code false positive.
-def enumerate_system_devices(  # dead: disable
-    interface_filter: Interface = None,
-) -> list:
-    """Iterate through all interfaces and locates available devices.
-
-    :param interface_filter: Interface enum to limit the search to a single interface type.
-    :return: A list of uosinterface objects.
-    """
-    system_devices = []
-    for interface in Interface:  # enum object
-        if not interface_filter or interface_filter == interface:
-            system_devices.extend(interface.value.enumerate_devices())
-        if interface_filter is not None:
-            break
-    return system_devices
-
-
-def get_device_definition(identity: str) -> Device | None:
-    """Look up the system config dictionary for the defined device mappings.
-
-    :param identity: String containing the lookup key of the device in the dictionary.
-    :return: Device Object or None if not found
-    """
-    if identity is not None and identity.lower() in DEVICES:
-        device = DEVICES[identity.lower()]
-    else:
-        device = None
-    return device
-
-
-ARDUINO_NANO_3 = Device(
+_ARDUINO_NANO_3 = Device(
     name="Arduino Nano 3",
     versions={},
     interfaces=[Interface.STUB, Interface.SERIAL],
@@ -148,7 +118,17 @@ ARDUINO_NANO_3 = Device(
     aux_params={"default_baudrate": 115200},
 )
 
-DEVICES = {
-    "hwid0": ARDUINO_NANO_3,
-    "arduino_nano": ARDUINO_NANO_3,
-}  # define aliases
+
+@dataclass(init=False, repr=False, frozen=True)
+class Devices:
+    """Names for supported hardware linking to the Device object used.
+
+    :cvar hwid_0: device: _ARDUINO_NANO_3
+    :cvar arduino_nano: device: _ARDUINO_NANO_3
+    :cvar arduino_uno: device: _ARDUINO_NANO_3
+    """
+
+    # Lookup constants linking devices to importable names
+    hwid_0: Device = _ARDUINO_NANO_3
+    arduino_nano: Device = _ARDUINO_NANO_3
+    arduino_uno: Device = _ARDUINO_NANO_3